requirements.txt
streamlit>=1.28.0
transformers>=4.35.0
datasets>=2.14.0
torch>=2.1.0
ollama>=0.1.0
requests>=2.31.0
python-docx>=1.1.0
tinydb>=4.8.0
accelerate>=0.24.0
peft>=0.7.0
bitsandbytes>=0.41.0
pyyaml>=6.0.0


